+++ src/ui/mod.rs
@@ -1,6 +1,7 @@
 /*! All user interface components, such as tabs, panels and dialogs.
 */
 pub mod bookmarks_tab;
+pub mod commands_tab;
 pub mod dialog;
 pub mod evolog_tab;
 pub mod files_tab;